+++ src/mathed/InsetMathCases.cpp
@@ -56,7 +56,7 @@ void InsetMathCases::draw(PainterInfo & pi, int x, int y) const
 	Changer dummy = pi.base.changeEnsureMath();
 	Dimension const dim = dimension(*pi.base.bv);
 	mathed_draw_deco(pi, x + 1, y - dim.ascent(), 6, dim.height(), from_ascii("{"));
-	int const t = mathed_deco_thickness(pi.base);
+	int const t = (int) mathed_deco_thickness(pi.base);
 	InsetMathGrid::draw(pi, x + t, y);
 }
